third_party/WebKit/LayoutTests/virtual/mediasession/media/mediasession/mediasession-constructor-expected.txt

CONSOLE WARNING: line 10: dummy log entry to avoid an empty -expected.txt in virtual/
This is a testharness.js-based test.
PASS MediaSession() constructor 
Harness: the test ran to completion.
